feat: conditional smoke test
Graphite follows GitHub required checks.
GitHub checks rely on target branch, but do not differentiate different
rules for source branch of PRs.
We will enable run-smoke-tests as a required GH check, but only run for
our gtmq_ and trying branches, and halting it by default for others.
This should act as a pass in the checks so that branches can be added to
the merge queue, and the merge queue draft PR will run the smoke tests