feat: modify update_database_track to use create_manifest style build (IN-2815) (#342)
## Description 
`update_database_track`  and `update_track` are very similar but have small differences in the tag being used and the track content structure. This PR refactors the script to only need `update_track` command and script to handle both, and adds a parameter `image_tag` to `update_database_track` job to be able to specify the -<arch> suffix. `create_manifest` is manually added in the .circleci of the database-cli repo instead of within the command.

## Related PRs
- https://github.com/voiceflow/database-cli/pull/887